feat: добавить полный заголовок ивента в десктопном разрешении
