+++ src/blocks/events/__title/_brief/events__title_brief.css
@@ -1,0 +1,3 @@
+.events__title_brief {
+  width: min-content;
+}
