+++ src/blocks/events/__title/events__title.css
@@ -1,26 +1,28 @@
 .events__title {
   font-family: "Roboto", sans-serif;
-  font-size: 30px;
-  font-weight: 400;
+  font-size: 24px;
   line-height: 36px;
+  font-weight: 500;
   color: rgba(191, 21, 35, 1);
-  padding: 16px 26px;
+  padding: 16px 8px;
   margin: 0;
   max-width: 480px;
   box-sizing: border-box;
   display: none;
   border: 1px solid white;
+  width: min-content;
+  text-align: center;
 }
 
 @media screen and (min-width: 425px) {
   .events__title {
-    padding: 16px 8px;
-    width: min-content;
-    font-size: 24px;
-    font-weight: 500;
+    width: 100%;
+    padding: 16px 26px;
+    font-weight: 400;
+    font-size: 30px;
     line-height: 29px;
     background-color: white;
-    text-align: center;
     display: inline-block;
+    text-align: left;
   }
 }
