+++ src/index.html
@@ -94,7 +94,7 @@
     </div>
 
     <template class="events-template">
-      <section class="events events_large events_title_sideway">
+      <section class="events events_large">
         <h2 class="events__title"></h2>
         <div class="events__container"></div>
       </section>
